requirements.txt
celery==5.1.2
Django==3.1
django-crispy-forms==1.12.0
django-storages==1.11.1
django-tinymce==3.3.0
Pillow==8.3.1

